Move utility methods to Utils class
