+++ php/class-delivery.php
@@ -17,6 +17,7 @@ use Cloudinary\UI\Component\HTML;
 use Cloudinary\Delivery\Bypass;
 use Cloudinary\Relate\Relationship;
 use WP_Post;
+use Cloudinary\Utils;
 
 /**
  * Plugin Delivery class.
@@ -1500,21 +1501,6 @@ class Delivery implements Setup {
 	}
 
 	/**
-	 * Get the path from a url.
-	 *
-	 * @param string $url The url.
-	 *
-	 * @return string
-	 */
-	public static function get_path_from_url( $url ) {
-		$content_url = content_url();
-		$path        = explode( Utils::clean_url( $content_url ), $url );
-		$path        = end( $path );
-
-		return $path;
-	}
-
-	/**
 	 * Check if the file type is allowed to be uploaded.
 	 *
 	 * @param string $ext The filetype extension.
@@ -1600,8 +1586,8 @@ class Delivery implements Setup {
 		$found                       = array();
 		$found[ $item['public_id'] ] = $item;
 		$url                         = Utils::clean_url( $content_url ) . $item['sized_url'];
-		$scaled                      = self::make_scaled_url( $url );
-		$descaled                    = self::descaled_url( $url );
+		$scaled                      = Utils::make_scaled_url( $url );
+		$descaled                    = Utils::descaled_url( $url );
 		$scaled_slashed              = addcslashes( $scaled, '/' );
 		$descaled_slashed            = addcslashes( $descaled, '/' );
 		$found[ $scaled ]            = $item;
@@ -1693,44 +1679,9 @@ class Delivery implements Setup {
 		) {
 			$sized                                = wp_basename( $url );
 			$url                                  = str_replace( '-' . $dash, '', $url );
-			$scaled                               = self::make_scaled_url( $url );
+			$scaled                               = Utils::make_scaled_url( $url );
 			$this->found_urls[ $url ][ $dash ]    = $sized;
 			$this->found_urls[ $scaled ][ $dash ] = $sized;
-		}
-
-		return $url;
-	}
-
-	/**
-	 * Make a scaled version.
-	 *
-	 * @param string $url The url to make scaled.
-	 *
-	 * @return string
-	 */
-	public static function make_scaled_url( $url ) {
-		$file = Utils::pathinfo( $url );
-		$dash = strrchr( $file['filename'], '-' );
-		if ( '-scaled' === $dash ) {
-			return $url;
-		}
-
-		return $file['dirname'] . '/' . $file['filename'] . '-scaled.' . $file['extension'];
-	}
-
-	/**
-	 * Make a descaled version.
-	 *
-	 * @param string $url The url to descaled.
-	 *
-	 * @return string
-	 */
-	public static function descaled_url( $url ) {
-		$file = Utils::pathinfo( $url );
-		$dash = strrchr( $file['filename'], '-' );
-		if ( '-scaled' === $dash ) {
-			$file['basename'] = str_replace( '-scaled.', '.', $file['basename'] );
-			$url              = $file['dirname'] . '/' . $file['basename'];
 		}
 
 		return $url;
@@ -1751,7 +1702,7 @@ class Delivery implements Setup {
 
 		// De-size.
 		$desized = array_unique( array_map( array( $this, 'maybe_unsize_url' ), $urls ) );
-		$scaled  = array_unique( array_map( array( $this, 'make_scaled_url' ), $desized ) );
+		$scaled  = array_unique( array_map( array( Utils::class, 'make_scaled_url' ), $desized ) );
 		$urls    = array_unique( array_merge( $desized, $scaled, $decoded ) );
 		$urls    = array_values( $urls ); // resets the index.
 
@@ -1768,7 +1719,7 @@ class Delivery implements Setup {
 			return; // Bail since theres nothing.
 		}
 
-		$paths = array_map( array( $this, 'get_path_from_url' ), $urls );
+		$paths = array_map( array( Utils::class, 'get_path_from_url' ), $urls );
 
 		$results = Utils::query_relations( $public_ids, $urls );
