+++ php/sync/class-storage.php
@@ -449,7 +449,7 @@ class Storage implements Notice {
 			$exists = $this->media->get_id_from_sync_key( $file_path );
 			if ( empty( $exists ) ) {
 				// Scaled size check.
-				$exists = $this->media->get_id_from_sync_key( Delivery::make_scaled_url( $file_path ) );
+				$exists = $this->media->get_id_from_sync_key( Utils::make_scaled_url( $file_path ) );
 				// Check for not synced items.
 				if ( empty( $exists ) && ! file_exists( $dir . DIRECTORY_SEPARATOR . $filename ) ) {
 					break;
